+++ hutool-log/src/main/java/com/xiaoleilu/hutool/log/dialect/console/ConsoleLogFactory.java
@@ -15,12 +15,12 @@ public class ConsoleLogFactory extends LogFactory {
 	}
 
 	@Override
-	public Log getLog(String name) {
+	public Log createLog(String name) {
 		return new ConsoleLog(name);
 	}
 
 	@Override
-	public Log getLog(Class<?> clazz) {
+	public Log createLog(Class<?> clazz) {
 		return new ConsoleLog(clazz);
 	}