+++ hutool-log/src/main/java/com/xiaoleilu/hutool/log/dialect/slf4j/Slf4jLogFactory.java
@@ -63,12 +63,12 @@ public class Slf4jLogFactory extends LogFactory {
 	}
 
 	@Override
-	public Log getLog(String name) {
+	public Log createLog(String name) {
 		return new Slf4jLog(name);
 	}
 
 	@Override
-	public Log getLog(Class<?> clazz) {
+	public Log createLog(Class<?> clazz) {
 		return new Slf4jLog(clazz);
 	}