+++ hutool-log/src/main/java/com/xiaoleilu/hutool/log/dialect/tinylog/TinyLogFactory.java
@@ -20,12 +20,12 @@ public class TinyLogFactory extends LogFactory {
 	}
 
 	@Override
-	public Log getLog(String name) {
+	public Log createLog(String name) {
 		return new TinyLog(name);
 	}
 
 	@Override
-	public Log getLog(Class<?> clazz) {
+	public Log createLog(Class<?> clazz) {
 		return new TinyLog(clazz);
 	}